+++ packages/@headlessui-vue/src/components/combobox/combobox.ts
@@ -387,10 +387,11 @@ export let Combobox = defineComponent({
     }
 
     // Handle outside click
-    useOutsideClick([inputRef, buttonRef, optionsRef], () => {
-      if (comboboxState.value !== ComboboxStates.Open) return
-      api.closeCombobox()
-    })
+    useOutsideClick(
+      [inputRef, buttonRef, optionsRef],
+      () => api.closeCombobox(),
+      computed(() => comboboxState.value === ComboboxStates.Open)
+    )
 
     watch([api.value, api.inputRef], () => api.syncInputValue(), {
       immediate: true,
@@ -429,7 +430,7 @@ export let Combobox = defineComponent({
     )
 
     return () => {
-      let { name, modelValue, disabled, ...incomingProps } = props
+      let { name, modelValue, disabled, ...theirProps } = props
       let slot = {
         open: comboboxState.value === ComboboxStates.Open,
         disabled,
@@ -456,10 +457,11 @@ export let Combobox = defineComponent({
             )
           : []),
         render({
-          props: {
+          theirProps: {
             ...attrs,
-            ...omit(incomingProps, ['nullable', 'multiple', 'onUpdate:modelValue', 'by']),
+            ...omit(theirProps, ['nullable', 'multiple', 'onUpdate:modelValue', 'by']),
           },
+          ourProps: {},
           slot,
           slots,
           attrs,
@@ -490,9 +492,11 @@ export let ComboboxLabel = defineComponent({
       }
 
       let ourProps = { id, ref: api.labelRef, onClick: handleClick }
+      let theirProps = props
 
       return render({
-        props: { ...props, ...ourProps },
+        ourProps,
+        theirProps,
         slot,
         attrs,
         slots,
@@ -591,9 +595,11 @@ export let ComboboxButton = defineComponent({
         onKeydown: handleKeydown,
         onClick: handleClick,
       }
+      let theirProps = props
 
       return render({
-        props: { ...props, ...ourProps },
+        ourProps,
+        theirProps,
         slot,
         attrs,
         slots,
@@ -629,7 +635,6 @@ export let ComboboxInput = defineComponent({
 
         case Keys.Backspace:
         case Keys.Delete:
-          if (api.comboboxState.value !== ComboboxStates.Open) return
           if (api.mode.value !== ValueMode.Single) return
           if (!api.nullable.value) return
 
@@ -728,7 +733,9 @@ export let ComboboxInput = defineComponent({
       let slot = { open: api.comboboxState.value === ComboboxStates.Open }
       let ourProps = {
         'aria-controls': api.optionsRef.value?.id,
-        'aria-expanded': api.disabled ? undefined : api.comboboxState.value === ComboboxStates.Open,
+        'aria-expanded': api.disabled.value
+          ? undefined
+          : api.comboboxState.value === ComboboxStates.Open,
         'aria-activedescendant':
           api.activeOptionIndex.value === null
             ? undefined
@@ -744,10 +751,11 @@ export let ComboboxInput = defineComponent({
         tabIndex: 0,
         ref: api.inputRef,
       }
-      let incomingProps = omit(props, ['displayValue'])
+      let theirProps = omit(props, ['displayValue'])
 
       return render({
-        props: { ...incomingProps, ...ourProps },
+        ourProps,
+        theirProps,
         slot,
         attrs,
         slots,
@@ -816,10 +824,11 @@ export let ComboboxOptions = defineComponent({
         ref: api.optionsRef,
         role: 'listbox',
       }
-      let incomingProps = omit(props, ['hold'])
+      let theirProps = omit(props, ['hold'])
 
       return render({
-        props: { ...incomingProps, ...ourProps },
+        ourProps,
+        theirProps,
         slot,
         attrs,
         slots,
@@ -882,7 +891,7 @@ export let ComboboxOption = defineComponent({
       api.selectOption(id)
       if (api.mode.value === ValueMode.Single) {
         api.closeCombobox()
-        nextTick(() => dom(api.inputRef)?.focus({ preventScroll: true }))
+        dom(api.inputRef)?.focus({ preventScroll: true })
       }
     }
 
@@ -926,8 +935,11 @@ export let ComboboxOption = defineComponent({
         onMouseleave: handleLeave,
       }
 
+      let theirProps = props
+
       return render({
-        props: { ...props, ...ourProps },
+        ourProps,
+        theirProps,
         slot,
         attrs,
         slots,