+++ packages/@headlessui-vue/src/utils/render.test.ts
@@ -1,5 +1,5 @@
-import { defineComponent, ComponentOptionsWithoutProps } from 'vue'
-import { render as testRender } from '../test-utils/vue-testing-library'
+import { defineComponent } from 'vue'
+import { createRenderTemplate } from '../test-utils/vue-testing-library'
 
 import { render } from './render'
 import { html } from '../test-utils/html'
@@ -9,25 +9,11 @@ let Dummy = defineComponent({
     as: { type: [Object, String], default: 'div' },
   },
   setup(props, { attrs, slots }) {
-    return () => render({ props, slots, attrs, slot: {}, name: 'Dummy' })
+    return () => render({ theirProps: props, ourProps: {}, slots, attrs, slot: {}, name: 'Dummy' })
   },
 })
 
-function renderTemplate(input: string | ComponentOptionsWithoutProps) {
-  let defaultComponents = { Dummy }
-
-  if (typeof input === 'string') {
-    return testRender(defineComponent({ template: input, components: defaultComponents }))
-  }
-
-  return testRender(
-    defineComponent(
-      Object.assign({}, input, {
-        components: { ...defaultComponents, ...input.components },
-      }) as Parameters<typeof defineComponent>[0]
-    )
-  )
-}
+const renderTemplate = createRenderTemplate({ Dummy })
 
 describe('Validation', () => {
   it('should error when using an as="template" with additional props', () => {
@@ -74,7 +60,8 @@ describe('Validation', () => {
         PassThrough(props, context) {
           props.as = props.as ?? 'template'
           return render({
-            props,
+            theirProps: props,
+            ourProps: {},
             attrs: context.attrs,
             slots: context.slots,
             slot: {},
@@ -92,5 +79,45 @@ describe('Validation', () => {
     })
 
     expect(document.getElementById('result')).toHaveClass('abc')
+  })
+
+  it('should allow use of <slot> as children', () => {
+    renderTemplate({
+      template: html`
+        <ExampleOuter>
+          <div id="result">Some Content</div>
+        </ExampleOuter>
+      `,
+
+      components: {
+        ExampleOuter: defineComponent({
+          template: html`
+            <ExampleInner>
+              <slot />
+            </ExampleInner>
+          `,
+
+          components: {
+            ExampleInner: defineComponent({
+              components: { Dummy },
+
+              template: html`
+                <Dummy as="template" class="foo" data-test="123">
+                  <Dummy as="template" class="bar" data-test="345">
+                    <slot />
+                  </Dummy>
+                </Dummy>
+              `,
+            }),
+          },
+        }),
+      },
+    })
+
+    expect(document.getElementById('result')).toHaveClass('foo')
+    expect(document.getElementById('result')).toHaveClass('bar')
+
+    // TODO: Is this the expected behavior? Should it actually be `345`?
+    expect(document.getElementById('result')).toHaveAttribute('data-test', '123')
   })
 })